+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+- Removed
+  - Support for the (experimental) OTG format download
+
 ## [2.3.3] - 2021-11-04
 
 - Fixed